chore: marking tests as @MainActor rather than class
adding deprecation warnings to optional property protocols and putting in own file
