+++ GDSCommon-Demo/GDSCommon-DemoTests/GDSErrorViewControllerTests.swift
@@ -147,6 +147,7 @@ private struct TestViewModelWithTertiary: GDSErrorViewModelV2, GDSErrorViewModel
 }
 
 extension GDSErrorViewControllerTests {
+    @MainActor
     func test_labelContents() throws {
         XCTAssertNotNil(try sut.errorImage)
         XCTAssertEqual(try sut.errorTitleLabel.text, "Error screen title")
@@ -159,12 +160,14 @@ extension GDSErrorViewControllerTests {
         XCTAssertTrue(try sut.errorTertiaryButton.isHidden)
     }
     
+    @MainActor
     func test_primaryButtonAction() throws {
         XCTAssertFalse(primaryButton)
         try sut.errorPrimaryButton.sendActions(for: .touchUpInside)
         XCTAssertTrue(primaryButton)
     }
     
+    @MainActor
     func test_secondaryButtonAction() throws {
         XCTAssertFalse(secondaryButton)
         try sut.errorSecondaryButton.sendActions(for: .touchUpInside)
@@ -205,7 +208,8 @@ extension GDSErrorViewControllerTests {
         XCTAssertEqual(try sut.errorTertiaryButton.title(for: .normal), "Error tertiary button title")
     }
     
-    func test_didAppear() throws {
+    @MainActor
+    func test_didAppear() {
         XCTAssertFalse(viewDidAppear)
         sut.beginAppearanceTransition(true, animated: false)
         sut.viewDidAppear(false)
