+++ GDSCommon-Demo/GDSCommon-DemoTests/GDSLoadingViewControllerTests.swift
@@ -8,7 +8,8 @@ final class GDSLoadingViewControllerTests: XCTestCase {
     var sut: GDSLoadingViewController!
     var didAppear: Bool!
     var didDismiss: Bool!
-
+    
+    @MainActor
     override func setUp() {
         super.setUp()
         
@@ -23,7 +24,7 @@ final class GDSLoadingViewControllerTests: XCTestCase {
         
         sut = .init(viewModel: viewModel)
     }
-
+    
     override func tearDown() {
         viewModel = nil
         sut = nil
@@ -31,7 +32,8 @@ final class GDSLoadingViewControllerTests: XCTestCase {
         didDismiss = nil
         super.tearDown()
     }
-
+    
+    @MainActor
     func testDidAppear() throws {
         XCTAssertFalse(didAppear)
         sut.beginAppearanceTransition(true, animated: false)
@@ -39,7 +41,8 @@ final class GDSLoadingViewControllerTests: XCTestCase {
         sut.viewWillAppear(false)
         XCTAssertTrue(didAppear)
     }
-
+    
+    @MainActor
     func test_didDismiss() {
         XCTAssertFalse(didAppear)
         sut.beginAppearanceTransition(true, animated: false)
@@ -50,11 +53,11 @@ final class GDSLoadingViewControllerTests: XCTestCase {
         _ = sut.navigationItem.rightBarButtonItem?.target?.perform(sut.navigationItem.rightBarButtonItem?.action)
         XCTAssertTrue(didDismiss)
     }
-
+    
+    @MainActor
     func test_loadingLabelIsSet() {
         XCTAssertEqual(try sut.loadingLabel.text, "Test key")
     }
-
 }
 
 extension GDSLoadingViewController {
