+++ GDSCommon-Demo/GDSCommon-DemoTests/ResultsViewControllerTests.swift
@@ -57,6 +57,7 @@ private struct TestViewModel: ResultsViewModel, BaseViewModel {
 }
 
 extension ResultsViewControllerTests {
+    @MainActor
     func testDidAppear() {
         XCTAssertFalse(screenDidAppear)
         sut.beginAppearanceTransition(true, animated: false)
@@ -82,6 +83,7 @@ extension ResultsViewControllerTests {
         XCTAssertEqual(try sut.resultsButton.title(for: .normal), "Results button title")
     }
     
+    @MainActor
     func test_buttonAction() throws {
         XCTAssertFalse(screenDidDismiss)
         try sut.resultsButton.sendActions(for: .touchUpInside)
